edit listner code
