--- test/NGSILD/listener/configOrion.json
@@ -1,9 +1,0 @@
-{
-        "my_ip": "http://180.179.214.202",
-        "update_broker_url": "http://180.179.214.202:1026",
-        "subscribe_broker_url": "http://180.179.214.202:1026",
-        "type": "latency",
-        "id":"urn:ngsi-ld:latency:A102"
-}
-~
-
