--- test/NGSILD/listener/configScorpio.json
@@ -1,8 +1,0 @@
-{
-        "my_ip": "http://180.179.214.202",
-        "update_broker_url": "http://180.179.214.202:9090",
-        "subscribe_broker_url": "http://180.179.214.202:9090",
-        "type": "latency",
-        "id":"urn:ngsi-ld:latency:A102"
-}
-
