+++ test/NGSILD/subscriptionLatency/configFog.json
@@ -1,0 +1,7 @@
+{
+	"my_ip": "http://180.179.214.202",	
+	"update_broker_url": "http://180.179.214.202:8070",	
+	"subscribe_broker_url": "http://180.179.214.202:8070",
+	"type": "latency",
+	"id":"urn:ngsi-ld:latency:A102"
+}
